requirements.txt
txasgiresource>=2.0.0
twisted>=17.1.0
service_identity
channels>=2.0.0
Django>=2.0.0,<2.99.99
pytz>=2016.7
django-environ
django-cors-headers>=1.2.2
django-filter>=2.0.0
django-picklefield>=0.3.2
djangorestframework>=3.4.7
djangorestframework-filters>=1.0.0.dev0
djangorestframework-jsonapi>=2.4.0
jsonfield>=1.0.3
scandir>=1.3
requests>=2.11.1
rarfile>=2.7
json-rpc>=1.10.3
Whoosh>=2.7.4
guessit>=2.0.4
pillow
tvdb-api>=2.0,<2.999999
imdbparser
malparser
thomas>=2.2.1
apscheduler
django-cache-url
dj-database-url
wrapt
wampyre>=1.0.1
certifi
unplugged>=0.1.1
